// src/types/writer-persistence-authorization-reconsideration-remediation.ts
export type WriterPersistenceAuthorizationReconsiderationRemediationMode =
  "persistence_adapter_implementation_authorization_reconsideration_remediation_plan_only";

export type WriterPersistenceAuthorizationReconsiderationRemediationStatus =
  | "external_remediation_required"
  | "manual_review_required";

export type WriterPersistenceAuthorizationReconsiderationRemediationItem = {
  id: string;
  category: WriterPersistenceAuthorizationRemediationCategory;
  title: string;
  status: WriterPersistenceAuthorizationReconsiderationRemediationStatus;
  owner: WriterPersistenceAuthorizationRemediationOwner;
  sourceNoGoStatus: WriterPersistenceAuthorizationReconsiderationNoGoStatus;
  sourceNoGoItemIds: string[];
  sourcePreflightItemIds: string[];
  sourceReviewItemIds: string[];
  sourceRemediationItemIds: string[];
  sourceRefs: string[];
  blockerSummary: string;
  remediationObjective: string;
  externalActions: string[];
  safeEvidenceRequirements: string[];
  verificationSteps: string[];
  acceptanceCriteria: string[];
  residualRisks: string[];
  redactionRules: string[];
  forbiddenActions: string[];
  nonExecutionClauses: string[];
  exitCriteria: string[];
  nextReviewGate: string;
};

export type WriterPersistenceAuthorizationReconsiderationRemediationRuntimeFlags =
  {
    allRuntimeEffectsBlocked: true;
    wouldAcceptReconsiderationRemediation: false;
    wouldRecordReconsiderationRemediationEvidence: false;
    wouldMarkReconsiderationBlockerResolved: false;
    wouldCreateReconsiderationRemediationTicket: false;
    wouldAcceptReconsiderationNoGo: false;
    wouldRecordReconsiderationNoGo: false;
    wouldDenyImplementationAuthorizationFromReconsideration: false;
    wouldPromoteToReconsiderationRemediation: false;
    wouldAcceptReconsiderationPreflight: false;
    wouldRecordReconsiderationPreflight: false;
    wouldMarkReconsiderationReady: false;
    wouldStartAuthorizationReconsideration: false;
    wouldAcceptRemediationReviewNoGo: false;
    wouldRecordRemediationReviewNoGo: false;
    wouldAcceptRemediationReview: false;
    wouldRecordRemediationReview: false;
    wouldStoreRemediationReviewEvidence: false;
    wouldMarkExternalRemediationReviewed: false;
    wouldAcceptExternalRemediationState: false;
    wouldAcceptExternalApprovalArchive: false;
    wouldStoreApprovalArtifact: false;
    wouldUploadApprovalArtifact: false;
    wouldReadExternalArtifact: false;
    wouldHashExternalArtifact: false;
    wouldPersistArchiveIndex: false;
    wouldMarkArchiveComplete: false;
    wouldCreateAuthorizationRecord: false;
    wouldRecordAuthorizationDecision: false;
    wouldRecordAuthorizationNoGoDecision: false;
    wouldAcceptAuthorizationNoGoDecision: false;
    wouldDenyImplementationAuthorization: false;
    wouldGrantImplementationAuthorization: false;
    wouldRecordHumanDecision: false;
    wouldAcceptHumanDecision: false;
    wouldStoreDecisionArtifact: false;
    wouldAcceptReleaseNoGo: false;
    wouldRecordGoDecision: false;
    wouldGrantReleaseApproval: false;
    wouldEnableFeatureFlag: false;
    wouldDeployCode: false;
    wouldRunProductionWriter: false;
    wouldCollectSignature: false;
    wouldRecordOwnerApproval: false;
    wouldGrantImplementationApproval: false;
    wouldCreateApprovalRecord: false;
    wouldAcceptPatchReview: false;
    wouldReviewRealPatch: false;
    wouldAcceptPatch: false;
    wouldGeneratePatch: false;
    wouldApplyPatch: false;
    wouldModifyFiles: false;
    wouldCreateFiles: false;
    wouldDeleteFiles: false;
    wouldRunGitCommand: false;
    wouldCreateBranch: false;
    wouldCheckoutBranch: false;
    wouldCreatePullRequest: false;
    wouldCreateTestFiles: false;
    wouldRunAutomatedTests: false;
    wouldCreateImplementationPlan: false;
    wouldCreateImplementationBranch: false;
    wouldCreateAdapterCode: false;
    wouldImportRealWriterImplementation: false;
    wouldRunTransaction: false;
    wouldCreateServiceRoleClient: false;
    wouldReadServiceRoleSecret: false;
    wouldPersistEvidence: false;
    wouldStoreRawPayload: false;
    wouldStoreSecrets: false;
    wouldWriteRows: false;
    wouldWriteAuditRows: false;
    wouldReserveIdempotencyKeys: false;
    wouldWriteIdempotencyRows: false;
    wouldWriteCompensationRows: false;
    wouldCreateMigrationFile: false;
    wouldApplyMigration: false;
    wouldCreateTables: false;
    wouldEnableWriters: false;
    wouldCallAi: false;
    wouldCallStripe: false;
    wouldUnlockReports: false;
  };

export type WriterPersistenceAuthorizationReconsiderationRemediationPayload =
  WriterPersistenceAuthorizationReconsiderationRemediationRuntimeFlags & {
    safeMode: true;
    readOnly: true;
    reconsiderationRemediationMode: WriterPersistenceAuthorizationReconsiderationRemediationMode;
    sourceReconsiderationNoGoMode: WriterPersistenceAuthorizationReconsiderationNoGoMode;
    checkedAt: string;
    remediationItemCount: number;
    externalRemediationRequiredCount: number;
    manualReviewRequiredCount: number;
    externalActionCount: number;
    safeEvidenceRequirementCount: number;
    verificationStepCount: number;
    acceptanceCriteriaCount: number;
    residualRiskCount: number;
    redactionRuleCount: number;
    forbiddenActionCount: number;
    exitCriteriaCount: number;
    sourceNoGoItemCount: number;
    sourceNoGoCount: number;
    sourceManualReviewBlockedCount: number;
    sourceReconsiderationStillBlockedCount: number;
    reconsiderationRemediationPlanReady: true;
    reconsiderationRemediationPlanOnly: true;
    sourceReconsiderationNoGoPacketReady: true;
    sourceReconsiderationNoGoPacketOnly: true;
    sourcePreflightChecklistReady: true;
    sourcePreflightChecklistOnly: true;
    sourceReviewNoGoPacketReady: true;
    sourceReviewNoGoPacketOnly: true;
    sourceReleaseStillBlocked: true;
    preflightPassed: false;
    preflightAccepted: false;
    preflightRecorded: false;
    reconsiderationEligible: false;
    reconsiderationNoGoAccepted: false;
    reconsiderationNoGoRecorded: false;
    reconsiderationRemediationAccepted: false;
    reconsiderationRemediationRecorded: false;
    implementationAuthorizationReconsiderationReady: false;
    implementationAuthorizationRemediationAccepted: false;
    implementationAuthorizationDecisionReady: false;
    implementationAuthorizationDecisionRecorded: false;
    implementationAuthorizationNoGoAccepted: false;
    implementationAuthorizationDenied: false;
    implementationAuthorizationGranted: false;
    implementationAuthorized: false;
    authorizationDecisionRecorded: false;
    authorizationArtifactStored: false;
    externalRemediationStatesAccepted: false;
    remediationReviewAccepted: false;
    remediationReviewComplete: false;
    remediationReviewNoGoAccepted: false;
    remediationReviewNoGoRecorded: false;
    externalApprovalArchiveAccepted: false;
    archiveCompletenessAccepted: false;
    implementationApprovalGranted: false;
    implementationBranchApproved: false;
    implementationPlanApproved: false;
    readyToApplyPatch: false;
    readyToCreateImplementationBranch: false;
    readyForAdapterImplementation: false;
    readyForReleaseExecution: false;
    adapterImplemented: false;
    adapterImplementationApproved: false;
    adapterImplementationAllowed: false;
    implementationReviewComplete: false;
    allOwnerApprovalsComplete: false;
    allBlockingEvidenceReady: false;
    blockedCodes: string[];
    remediationPlanRules: string[];
    remediationReviewRules: string[];
    sourceBlockedCodes: string[];
    remediationItems: WriterPersistenceAuthorizationReconsiderationRemediationItem[];
  };

export type WriterPersistenceAuthorizationReconsiderationRemediationProbeResult =
  WriterPersistenceAuthorizationReconsiderationRemediationRuntimeFlags & {
    safeMode: true;
    readOnly: true;
    blocked: true;
    reconsiderationRemediationMode: WriterPersistenceAuthorizationReconsiderationRemediationMode;
    itemId?: string;
    itemTitle?: string;
    itemStatus?: WriterPersistenceAuthorizationReconsiderationRemediationStatus;
    summary: string;
    reconsiderationRemediationPlanOnly: true;
    sourceReconsiderationNoGoPacketOnly: true;
    sourcePreflightChecklistOnly: true;
    sourceReviewNoGoPacketOnly: true;
    sourceReleaseStillBlocked: true;
    preflightAccepted: false;
    preflightRecorded: false;
    reconsiderationEligible: false;
    reconsiderationNoGoAccepted: false;
    reconsiderationNoGoRecorded: false;
    reconsiderationRemediationAccepted: false;
    reconsiderationRemediationRecorded: false;
    implementationAuthorizationReconsiderationReady: false;
    implementationAuthorizationGranted: false;
    implementationAuthorized: false;
    authorizationDecisionRecorded: false;
    authorizationArtifactStored: false;
    externalRemediationStatesAccepted: false;
    remediationReviewAccepted: false;
    remediationReviewComplete: false;
    remediationReviewNoGoAccepted: false;
    remediationReviewNoGoRecorded: false;
    externalApprovalArchiveAccepted: false;
    archiveCompletenessAccepted: false;
    readyToCreateImplementationBranch: false;
    readyForAdapterImplementation: false;
    readyForReleaseExecution: false;
    adapterImplemented: false;
    adapterImplementationApproved: false;
    adapterImplementationAllowed: false;
    allOwnerApprovalsComplete: false;
    allBlockingEvidenceReady: false;
    blockedCodes: string[];
    remediationItems: WriterPersistenceAuthorizationReconsiderationRemediationItem[];
  };